+++ screens/ProfileScreen.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { View, Text } from 'react-native';
+
+const ProfileScreen = ({navigation}) => {
+    return (
+        <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Profile Screen</Text>
+    </View>
+    );
+}
+export default ProfileScreen;
